Fix null reference exception on CitirocUI load
The form was crashing on start on a Windows 10
machine, because a call to the form's LocationChanged event was made at the start, and this form was calling the
SendWindowPositionEvent. This event is used to change the position of the monitor form, and with a non-existing monitor form, the event was NULL -- and this was triggering a null pointer exception.

Interestingly, this was not happening on a Windows 7
machine.