Save DAQ data to file from ProtoCubesSerial
Also _actually_ set the default DAQ time to 5s, for debug purposes, by setting the default value of the text box to 5s. (I was previously assigning the default value to the string member inside DataAcquisition.cs,
which is assigned the setting in the data acquisition
time text box when DAQ starts. So it is this setting in
the text box which is actually used, not the default
setting you assign to the string during the variable's
declaration.)