Checking for valid path now within try-catch statement
In the previous implementation, the validPath() function
could either return a null file in case of invalid file name,
such as one containing forward slashes, or throw an
exception in case a colon existed in the path. The latter
case was uncovered; this is fixed with this commit, by
moving the call to validPath() in a try-catch statement
and throwing an exception in case the function returns
false.