Check for valid data file path early
Before, characters were sent over the comm line to start
DAQ despite the fact that the DAQ process wasn't
actually starting on the CitirocUI end. This could
happen if a call to validPath() yielded a wrong file
path.

The fix to this is to place the checking of valid data
path at the start of the button_startAcquisition_Click()
function.