Change DAQ output file naming
1. Added the Swedish locale for converting the date-time string. The
Swedish locale allows for file names to appear ordered by date in a file
explorer window (or when using 'ls'), since the date in Swedish locale is
according to ISO-8601.

2. Start the file name with dataCITI, as it was previously done by Weeroc.
So, the DAQ file name will be "dataCITI_<date>.dat"

Also added an unrelated, small formatting change in the message box
shown in case of validPath() failing.