Fix (again) point of DAQ output file writing
This time, the checking for number of DAQ bytes retrieved vs. the
expected number of bytes (in this case, the DAQ data array length)
was moved right after  data bytes are copied into the DAQ byte array.
It is probably the best location to make this check -- and it works.

Also made the help label text more helpful in DataAcquisition.cs.