Implement TM data readout in ProtoCubesSerial (WIP)
The dummy functions have been moved from Connect.cs to
ProtoCubesSerial.cs. Some modifications there have also been made to
remove the code previously implemented to cover debug data being
written by the Arduino to the serial port, along with some other fixes.

The code is still WIP, it needs the implementation of (I think) an event
triggered by ProtoCubesSerial when data is ready, an event that CitirocUI
subscribes to.

The current code does compile and run, it just creates an exception
when ProtoCubesSerial tries to write directly to the monitor form.