Implement inefficient and incomplete DataReadyEvent (WIP)
I started implementing a DataReadyEvent to retrieve the data from a
housekeeping run. It is highly inefficient, as it copies over the multiple
properties needed to write into the monitor form into a class, which
needs its own private members with associated properties. In order to
write these members from the ProtoCubesSerial class, the members
need to be redefined inside ProtoCubesSerial AND additionally they are
also defined inside frmMontir. Hence the inefficiency.

Quitting while ahead, regrouping and reimplementing on Monday (or
a bit later on) the following:

The DataReadyEventArgs will simply contain the data buffer and the
data type present in the data buffer (probably along with the number
of bytes stored). Then, the subscribing class (CitirocUI) will decode
the data buffer and present it to frmMonitor.