Adjust total DAQ time according to individual DAQ time
The total DAQ time will be set to the individual DAQ time if a time lower
than the individual DAQ time is set.

This happens only in the case of Proto-CUBES being selected as the
connection medium.

A small background change was to remove the naming of the
selectedConnectionMode member -- it does not have an underscore
at preceding it, this makes it more "C#-y".