Avoid exceptions and apply formatting to total DAQ time setting
This commit avoids exceptions when the user leaves the total DAQ time
setting with unfilled details by setting all unset fields to zeroes. The changes
in this commit also avoid unclear formatting such as "00:1_:01", or
"00"1_:5_", which would translate after conversion to Int32 to"
0h:1m:1s" and "0h:1m:5s", respectively. The formatting ensured is
two digits for each of the fields (hours, minutes, seconds).