Move DAQ data file storage to DataAcquisition.cs
This is done through the following steps:

1. Adding a call to a DataReadyEvent in ProtoCubesSerial's
DataReceived COM port event

2. Adding a handler to the DataReadyEvent in the
Citiroc class [in the Citiroc_Load() method, inside
Citiroc.cs]

3. Implementing the handler inside DataAcquisition.cs

This commit also adds regions to DataAcquisition.cs and
moves methods to their appropriate regions.

DAQ data storage does not work properly yet, some
timing issues are to be handled on the Arduino side,
then the CitirocUi has to be synchronized to the changes
on the Arduino.