Preclude exception when HVPS turned off
The CUBES Monitor part of the CitirocUI was throwing an exception when
the HVPS was turned off. This was because the HVPS sends all zeroes when
turned off, and the GetString() method used to convert the received bytes
into a string was converting the "zero" bytes into null-terminators. These
null-terminators were then causing an exception when using the
Convert.ToDouble() method.

This is fixed in a very poor-man style, by simply checking that the
string is comprised of four null-terminators. If it is, turn it into a series
of four "zero" characters.