Implement configuring HVPS voltage
The configuration is done via the HVPS groupbox in the Connect tab. When
the button in this group box is clicked, two bytes are sent to the Arduino
via the CMD_SEND_HVPS_TMP_VOLT command. The Arduino then sends
the bytes over MSP, which are converted to the HVPS's HBV command
by the Cortex-M3 on the SmartFusion2.