Add readout of HVPS command count
These extra TM obtained from CUBES is displayed in the new text boxes added
under the CUBES monitor.  The ProtoCubesSerial class has also been updated
to reflect the extra six bytes obtained from CUBES as part of the HVPS
command counters.

The text boxes used to display the counters show in the help label some
text to explain what they contain. The MouseHover and MouseLeave
events are used to display to and clear the help label. The MouseLeave
event has also been added to the HVPS reset check-box, with the same
action as for the HVPS counter text boxes -- clear the help label.