Add Clear SD Card button
The button is used to send the "Delete SD Card Files" command to the
Arduino. The command is added in ProtoCubesSerial.cs. The Clear SD Card
button's event handler in Citiroc.cs is used to send the command to
Proto-CUBES. A MessageBox is used to ensure the user is sure he/she
actually wants to send the command to delete the files.