Use protoCubes.NumBins instead of UI combo box value
This is a value which is not UI-thread specific, hence it avoids the
usual .NET thread errors. Such an error resulted in the previous lack
of functionality of the DAQ, following the changes to add a BINCFG
param. to the DAQ configuration sent to CUBES.

The protoCubes.NumBins property is also set upon DAQ start and
not when DAQ ends, as it was done previously. This makes for a
more logical operation, in my view.

Additionally, the "num bins" label and combo-box have been
disabled during DAQ, to ensure no changes are made to it when
they shouldn't.