Remove comboBox_cubesDaqNumBins_SelectedIndexChanged
Actually, this event handler is not necessary, the HG and LG cut points
should not be changed by default. Changing the number of bins does
not change the top level of the ADC, which is what the HG and LG cut
points refer to. Instead, a lower number of bins results in data having
a coarser spread on the ADC range.