Properly use the number of bin setting in data display
Previously (for Proto-CUBES), the data display always assumed 2048
bins, as was the case for the Proto-CUBES 2019 flight. With the
rebinning implementation, the number of bins changes - and so do byte
offsets accordingly. This commit adjusts the byte offsets.

An issue still exists with the data display: The top channel of the ADC
is halved on every halving of the number of bins, e.g., with 2048 bins
the top channel is ~1500, for 1024 it is ~700, etc.  TBD if this is
because of CitirocUI or the Arduino or the actual rebinning implem. on
CUBES...