Fix data display and load data button with Num. Bins != 2048
The issue was using the protoCubes.NumBins drop-down to select the
number of bins to display. This is a user setting that is employed
before starting the DAQ.

When data is loaded from a file, however, the number of bins setting
in the file should be used, not the user setting, as the setting in the
file when DAQ was run may differ from the particular pre-DAQ setting at
a particular time.

Furthermore, the number of bins setting actually used by the
instrument should be used also when data is displayed post-DAQ.