Properly display re-binned data
By "properly" meaning in the right ADC range. With the way data
display worked previously, bin values were just displayed in sequence
as they were read. As there are fewer bins to display, this makes it
look like the ADC range becomes lower with lower number of bins. Not
so - data display should take into account the bin configuration.

This commit takes into account the bin configuration and results in
proper display of the data on the full ADC range.

In addition, some cleanup has been done in the background in this
commit.