Make command reply data length a member
Previously, a local variable was used in the DataReceived event handler
for the COM port used for communication with the instrument. The local
variable approach can make it that a REQ command changes the expected
data length of an on-going REQ command, if the on-going REQ command has
not finished processing.

This change, in collaboration with an upcoming implementation of a
command data buffer, should make the protocol more robust when it comes
to commands being properly executed.