Attempt to mitigate last data file issue
This is done by redefining the ProtoCubesSerial.currentCommand member
to ProtoCubesSerial.lastSentCommand, to (hopefully) ensure that the
higher-level DAQ code interprets the last bytes arriving from the
serial port as part of the last REQ_PAYLOAD command sent over to
Proto-CUBES. With the current implementation, what is observed is that
when all bytes corresponding to the previous REQ command arrive, the
setting of the command to ProtoCubesSerial.Command.None results in the
last bytes of the last REQ_PAYLOAD to not be interpreted any further.
This is due to the fact the event handler methods in DataAcquisition.cs
check for the command the event corresponds to, in order to properly
use the data arriving from Proto-CUBES.

The code in this commit is yet to be tested.