Use mask on calib. enable bit
There was a potential for a bug in the code with the way it was
previously implemented. When the calibration pulse enable checkbox was
checked, the code only applied the top-most bit to the most significant
byte sent to the hardware by doing `calibconfdata[0] = 0x80`. Bits
30..24 were all cleard to 0. This meant for a low enough frequency, the
potential existed for the frequency selected by the user to not be
applied to the hardware.

Using a mask to only set and clear the top-most bit according to the
state of the checkbox allows for using the whole 31 bits of the period
register.