Return of the checkbox
Re-added the checkbox for choosing to store on the NVM or directly add
to config RAM. Also added functionality that hides the checkbox when
proto-cubes is not selected. SelectNVMCitirocConf uses the 'load SC'
button when 'SendToNVM'-checkbox is checked.

Added class NVMConfigNumSelect.cs for input
of config number.

Functions will check if connected before prompting
user for input.

Changed name of UART commands to match the definition on
proto-cubes-obc.

Removed WeerocDotComLabel as clickable.