Post DataReadyEvent after timeout
With the changes in this commit, a ProtoCubesSerial.DataReadyEvent will
be posted after a timeout of 500 ms, even if not all bytes expected
have been received. This makes it possible among others to print
something in reply to the REQ_ID command, even if the REQ_CUBES_ID
on the CUBES board did not reply with any bytes.